requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-allauth==0.61.1
channels==4.0.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1 